refactored. missing unit tests
