+++ contract/processor.go
@@ -7,10 +7,8 @@ import (
 	"idp-cfs/client_git"
 	"idp-cfs/client_github"
 	"idp-cfs/util"
-	"io"
 	"os"
 	"path"
-	"path/filepath"
 	"strings"
 )
 
@@ -186,7 +184,7 @@ func (p *Processor) validateGoldenPath(dryRunMode bool) error {
 				gpPath = path.Join(p.GpClonePath, *p.Contract.GoldenPath.Path)
 			}
 
-			err = copyFilesFromGpToCode(gpPath, p.CodeClonePath)
+			err = util.CopyFilesDeep(gpPath, p.CodeClonePath)
 			if err != nil {
 				log.Error().Msgf("Failed to copy files from goldenpath to code repo: %v", err)
 				return err
@@ -202,44 +200,6 @@ func (p *Processor) validateGoldenPath(dryRunMode bool) error {
 		}
 
 		log.Info().Msg("Checked out branch successfully.")
-	}
-
-	return nil
-}
-
-func copyFilesFromGpToCode(srcDir string, destDir string) error {
-
-	err := filepath.Walk(srcDir, func(file string, info os.FileInfo, err error) error {
-		if !info.IsDir() {
-			srcFile, _ := os.Open(file)
-			defer func(srcFile *os.File) {
-				err := srcFile.Close()
-				if err != nil {
-					log.Error().Msgf("Failed to close the src file %s: %v", srcFile.Name(), err)
-				}
-			}(srcFile)
-
-			destFilePath := filepath.Join(destDir, info.Name())
-			destFile, _ := os.Create(destFilePath)
-			defer func(destFile *os.File) {
-				err := destFile.Close()
-				if err != nil {
-					log.Error().Msgf("Failed to close the src file %s: %v", srcFile.Name(), err)
-				}
-			}(destFile)
-
-			_, err := io.Copy(destFile, srcFile)
-			if err != nil {
-				log.Error().Msgf("Failed to copy the file from gp path to the new code path: %v", err)
-				return err
-			}
-		}
-		return nil
-	})
-
-	if err != nil {
-		log.Error().Msgf("Failed to walk the directory %s: %v", srcDir, err)
-		return err
 	}
 
 	return nil
